feat: to revision Activities Content Column
