+++ learn/lib/utils/activitiesList.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:learn/widgets/activitieCard.dart';
+import 'package:learn/widgets/activitiesWidgets/activitieCardStatic.dart';
 
-List<ActivitieCard> activitiesList = [
-   ActivitieCard(
+List<ActivitieCardStatic> activitiesList = [
+   ActivitieCardStatic(
     level: 1,
     title: "Orçamento para um passeio", 
     description:  "Continue onde parou",
@@ -11,7 +11,7 @@ List<ActivitieCard> activitiesList = [
     backgroundGradientColors: const  [Color(0XFF7A7FFF),Color(0xFF040862)],
     levelGradientColors: [Colors.white.withOpacity(0.3), Colors.white.withOpacity(0.01)],
   ),
-   ActivitieCard(
+   ActivitieCardStatic(
     level: 1,
     title: "Orçamento para um passeio", 
     description:  "Continue onde parou",
@@ -20,7 +20,7 @@ List<ActivitieCard> activitiesList = [
     backgroundGradientColors: const  [Color(0XFF1290A2),Color(0xFF82DA59)],
     levelGradientColors: [Colors.white.withOpacity(0.8), Colors.white.withOpacity(0.2)],
   ),
-  ActivitieCard(
+  ActivitieCardStatic(
     level: 2,
     title: "Orçamento para um passeio", 
     description:  "Continue onde parou",
